requirements.txt
langchain
langchain-groq
langchain-community
chromadb
streamlit
duckduckgo-search
requests
python-dotenv
langsmith